+++ cmd/nerdctl/ipfs/ipfs_build_linux_test.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package ipfs
 
 import (
 	"fmt"
