+++ cmd/nerdctl/ipfs/ipfs_registry_serve.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package ipfs
 
 import (
 	"github.com/spf13/cobra"
